+++ migrations/001-users.sql
@@ -1,23 +1,25 @@
 -- +migrate Up
-CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
-    phone_number VARCHAR(16) NOT NULL UNIQUE,
-    email VARCHAR(64),
-    password VARCHAR(256) NOT NULL,
-    profile VARCHAR(256),
-    first_name VARCHAR(128),
-    last_name VARCHAR(128),
-    display_name VARCHAR(128) NOT NULL,
-    gender INTEGER NOT NULL DEFAULT 0,
-    is_active BOOLEAN NOT NULL DEFAULT FALSE,
-    registered BOOLEAN NOT NULL DEFAULT FALSE,
-    deactivation_reason VARCHAR(256),
-    is_admin BOOLEAN NOT NULL DEFAULT FALSE,
-    otp_remaining_attempts INTEGER NOT NULL DEFAULT 0,
-    otp_code INTEGER,
-    otp_due_date TIMESTAMP,
-    is_superuser BOOLEAN NOT NULL DEFAULT FALSE,
-    created_at TIMESTAMP NOT NULL
+CREATE TABLE "users" (
+	"id" serial NOT NULL,
+	"phone_number" varchar(16) NOT NULL,
+	"email" varchar(64),
+	"password" varchar(256) NOT NULL,
+	"profile" varchar(256),
+	"first_name" varchar(128),
+	"last_name" varchar(128),
+	"display_name" varchar(128) NOT NULL,
+	-- 0 notdefined 1 male 2 female
+	"gender" int NOT NULL DEFAULT 0,
+	"is_active" boolean NOT NULL DEFAULT TRUE,
+	"registered" boolean NOT NULL DEFAULT FALSE,
+	"deactivation_reason" varchar(256),
+	"is_admin" boolean NOT NULL DEFAULT FALSE,
+	"otp_remaining_attempts" int NOT NULL DEFAULT 0,
+	"otp_code" int,
+	"otp_due_date" TIMESTAMPTZ,
+	"is_superuser" boolean NOT NULL DEFAULT FALSE,
+	"created_at" TIMESTAMPTZ NOT NULL,
+	PRIMARY KEY("id")
 );
 -- +migrate Down
-DROP TABLE users;
+DROP TABLE "users";