+++ migrations/002-tokens.sql
@@ -1,12 +1,12 @@
 -- +migrate Up
-CREATE TABLE tokens (
-    id SERIAL PRIMARY KEY,
-    token VARCHAR(256) NOT NULL,
-    is_refresh_token BOOLEAN NOT NULL DEFAULT FALSE,
-    user_id INTEGER NOT NULL,
-    expires_at TIMESTAMP NOT NULL,
-    created_at TIMESTAMP NOT NULL,
-    FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE
+CREATE TABLE "tokens" (
+	"id" serial NOT NULL,
+	"token" varchar(256) NOT NULL,
+	"is_refresh_token" boolean NOT NULL DEFAULT FALSE,
+	"user_id" int NOT NULL,
+	"expires_at" TIMESTAMPTZ NOT NULL,
+	"created_at" TIMESTAMPTZ,
+	PRIMARY KEY("id")
 );
 -- +migrate Down
-DROP TABLE tokens;
+DROP TABLE "tokens";