+++ migrations/009-users_groups_user_id_fk.sql
@@ -1,0 +1,8 @@
+-- +migrate Up
+ALTER TABLE "users_groups"
+ADD CONSTRAINT "users_groups_user_id_fk"
+FOREIGN KEY("user_id") REFERENCES "users"("id")
+ON UPDATE NO ACTION ON DELETE CASCADE;
+-- +migrate Down
+ALTER TABLE "tokens"
+DROP CONSTRAINT "users_groups_user_id_fk";