+++ routes/auth.go
@@ -1,0 +1,9 @@
+package routes
+
+import (
+	"github.com/kataras/iris/v12/core/router"
+)
+
+func AuthHTTP(app router.Party) {
+
+}