+++ routes/http.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/kataras/iris/v12"
+	"github.com/kataras/iris/v12/core/router"
 	"github.com/rs/cors"
 )
 
@@ -40,7 +41,7 @@ func addBasicErrorRoutes(app *iris.Application) {
 	app.OnErrorCode(iris.StatusNotFound, extra_middlewares.Translator, extra_middlewares.Panic, error_handlers.NotFound)
 }
 
-func HTTP(app *iris.Application) {
+func HTTP(app *iris.Application) router.Party {
 	addBasicMiddlewares(app)
 	addBasicErrorRoutes(app)
 
@@ -61,4 +62,6 @@ func HTTP(app *iris.Application) {
 		api := basicApi.Party("/api")
 		api.Get("/", handlers.Hello)
 	}
+
+	return basicApi
 }